+++ effect/effect.go
@@ -1,5 +1,5 @@
 /*Package effects provides the functionality to manipulate images to achieve various looks.*/
-package effects
+package effect
 
 import (
 	"image"
